update Delete confirmation to use in diff ways not just for delete.
